Move Pointer to types mod from memory
