+++ x/programs/rust/wasmlanche_sdk/src/types.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 use std::borrow::Cow;
 
-use crate::program::Program;
+use crate::{memory::Memory, program::Program};
 
 pub const ADDRESS_LEN: usize = 32;
 /// A struct that enforces a fixed length of 32 bytes which represents an address.
@@ -121,5 +121,58 @@ impl Argument for Program {
     }
     fn is_primitive(&self) -> bool {
         true
+    }
+}
+
+/// Represents a pointer to a block of memory allocated by the global allocator.
+/// A `Pointer` can be used directly as a param type on a `public` functions.
+///
+/// # Example
+/// ```
+/// use wasmlanche_sdk::types::Pointer;
+/// #[public]
+/// pub fn init( program: Program, name_ptr: Pointer, name_length: i64) -> bool {
+/// let nft_name = name_ptr.read(name_length as usize);
+/// ```
+///
+#[derive(Clone, Copy)]
+pub struct Pointer(*mut u8);
+
+impl Pointer {
+    pub fn new(ptr: *mut u8) -> Self {
+        Self(ptr)
+    }
+
+    #[must_use]
+    pub fn inner(&self) -> *mut u8 {
+        self.0
+    }
+
+    #[must_use]
+    /// Attempts return a copy of the bytes from a pointer created by the global allocator.
+    /// # Safety
+    /// `ptr` must be a pointer to a block of memory created using alloc.
+    /// `length` must be the length of the block of memory.
+    pub unsafe fn read(self, len: usize) -> Vec<u8> {
+        unsafe { Memory::new(self).range(len) }
+    }
+}
+
+impl From<i64> for Pointer {
+    fn from(v: i64) -> Self {
+        let ptr: *mut u8 = v as *mut u8;
+        Pointer(ptr)
+    }
+}
+
+impl From<Pointer> for *const u8 {
+    fn from(pointer: Pointer) -> Self {
+        pointer.0.cast_const()
+    }
+}
+
+impl From<Pointer> for *mut u8 {
+    fn from(pointer: Pointer) -> Self {
+        pointer.0
     }
 }
